pyproject.toml
[tool.poetry]
name = "mlflow-example"
version = "0.1.0"
description = ""
authors = ["Pei Sabrina Xu <pxu@mdsol.com>"]
readme = "README.md"
packages = [{include = "mlflow_example"}]

[tool.poetry.dependencies]
python = "3.11.2"
scikit-learn = "^1.2.2"
mlflow = "^2.4.1"
pip = "^24.0"
xgboost = "^2.0.3"
optuna = "^3.5.0"
seaborn = "^0.13.2"
plotly = "^5.18.0"
nbformat = "^5.9.2"
hydra-core = "^1.3.2"
jupytext = "^1.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
